feat: implement tracks K5 and K6 (Operational Transparency & Temporal Risk Precision)
K5:
- Added 'config view' command to show resolved project state.
- Implemented '--limit' and '--offset' pagination for 'ledger audit' and 'ledger search'.
- Added 'RECENT COMMITTED ENTRIES' section to global audit view.

K6:
- Added '--commits', '--days', and '--since' filters to 'hotspots' command.
- Implemented exponential decay weighting for hotspot scores (recent commits matter more).
- Updated Hotspot model and display logic to support weighted frequency (float).
- Fixed short flag collision in hotspots CLI (-s now specifically for --semantic).

Refactored calculate_hotspots to accept decay_half_life as a parameter for better testability and layout-awareness.

+++ src/impact/packet.rs
@@ -299,7 +299,7 @@ pub struct Hotspot {
     #[serde(default)]
     pub display_score: f32,
     pub complexity: i32,
-    pub frequency: usize,
+    pub frequency: f64,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub centrality: Option<usize>,
 }
